+++ Wire-iOS Tests/ReferenceImages/CallInfoRootViewControllerSnapshotTests/testOneToOneIncomingAudioRinging.iPhone-4_0_Inch.png
