+++ Wire-iOS Tests/ReferenceImages/CallInfoRootViewControllerSnapshotTests/testOneToOneIncomingAudioRinging.iPhone-4_7_Inch.png
