+++ Wire-iOS Tests/ReferenceImages/CallInfoRootViewControllerSnapshotTests/testOneToOneIncomingAudioRinging.iPhone-5_5_Inch.png
