+++ Wire-iOS Tests/ReferenceImages/CallInfoRootViewControllerSnapshotTests/testOneToOneIncomingAudioRinging.iPhone-5_8_Inch.png
