+++ Wire-iOS Tests/ReferenceImages/CallInfoRootViewControllerSnapshotTests/testOneToOneIncomingAudioRinging.iPhone-6_5_Inch.png
